feat: Add ability to toggle the theme from the drawer
