+++ example/main.js
@@ -2,72 +2,43 @@
 
 import Expo from 'expo';
 import React, { Component } from 'react';
-import PropTypes from 'prop-types';
-import { View, StyleSheet, Platform, StatusBar } from 'react-native';
+import { StatusBar } from 'react-native';
 import {
-  DrawerItem,
-  DrawerSection,
   Provider as PaperProvider,
-  withTheme,
   DarkTheme,
+  DefaultTheme,
 } from 'react-native-paper';
 import { DrawerNavigator } from 'react-navigation';
 import RootNavigator from './src/RootNavigator';
+import DrawerItems from './DrawerItems';
 
 StatusBar.setBarStyle('light-content');
 
-const DrawerItemsData = [
-  { label: 'Inbox', icon: 'inbox', key: 0 },
-  { label: 'Starred', icon: 'star', key: 1 },
-  { label: 'Sent mail', icon: 'send', key: 2 },
-  { label: 'A very long title that will be truncated', icon: 'delete', key: 3 },
-  { label: 'No Icon', key: 4 },
-];
+const App = DrawerNavigator(
+  { Home: { screen: RootNavigator } },
+  {
+    contentComponent: ({ screenProps }) =>
+      <DrawerItems toggleTheme={screenProps.toggleTheme} />,
+  }
+);
 
-class DrawerItems extends Component {
-  static propTypes = {
-    theme: PropTypes.object.isRequired,
-  };
+class PaperExample extends Component {
   state = {
-    open: false,
-    drawerItemIndex: 0,
+    theme: DarkTheme,
   };
-  _setDrawerItem = index => this.setState({ drawerItemIndex: index });
+
+  _toggleTheme = () =>
+    this.setState({
+      theme: this.state.theme === DarkTheme ? DefaultTheme : DarkTheme,
+    });
 
   render() {
-    const { theme: { colors: { paper } } } = this.props;
     return (
-      <View style={[styles.drawerContent, { backgroundColor: paper }]}>
-        <DrawerSection label="Subheader">
-          {DrawerItemsData.map((props, index) =>
-            <DrawerItem
-              {...props}
-              key={props.key}
-              active={this.state.drawerItemIndex === index}
-              onPress={() => this._setDrawerItem(index)}
-            />
-          )}
-        </DrawerSection>
-      </View>
+      <PaperProvider theme={this.state.theme}>
+        <App screenProps={{ toggleTheme: this._toggleTheme }} />
+      </PaperProvider>
     );
   }
 }
-const ThemedDrawerItems = withTheme(DrawerItems);
 
-const App = DrawerNavigator(
-  { Home: { screen: RootNavigator } },
-  { contentComponent: () => <ThemedDrawerItems /> }
-);
-
-const styles = StyleSheet.create({
-  drawerContent: {
-    flex: 1,
-    paddingTop: Platform.OS === 'android' ? 25 : 22,
-  },
-});
-
-Expo.registerRootComponent(() =>
-  <PaperProvider theme={DarkTheme}>
-    <App />
-  </PaperProvider>
-);
+Expo.registerRootComponent(PaperExample);
